tools/requirements.txt
PyQt6==6.4.2
pyserial==3.5
